+++ gateway/src/resty/oidc/discovery.lua
@@ -8,10 +8,16 @@ local resty_env = require 'resty.env'
 local Mime = require 'resty.mime'
 local cjson = require('cjson')
 local jwk = require('resty.oidc.jwk')
+local resty_lrucache = require "resty.lrucache"
+
 
 local oidc_log_level = ngx[string.upper(resty_env.value('APICAST_OIDC_LOG_LEVEL') or 'err')] or ngx.ERR
 
-local _M = { }
+local cache_max_size = 100
+
+local _M = {
+  _TYPE = "oidc_discovery"
+}
 
 local mt = { __index = _M }
 
@@ -40,10 +46,41 @@ function _M.new(http_backend)
             ssl = { verify = resty_env.enabled('OPENSSL_VERIFY') }
         }
     }
+    return _M.new_with_http_client(http_client)
+end
 
-    local self = { http_client = http_client }
+function _M.new_with_http_client(http_client)
+    local self = setmetatable({ http_client = http_client }, mt)
+    self:init_cache()
+    return self
+end
 
-    return setmetatable(self, mt)
+
+function _M:init_cache()
+  self.cache = resty_lrucache.new(cache_max_size)
+  return self
+end
+
+
+--- Fetch and return OIDC configuration from cache if it exists.
+function _M:issuer_in_cache(issuer)
+  if not self.cache then
+    return nil
+  end
+
+  return self.cache:get(issuer)
+end
+
+--- Insert the OIDC configuration for the issuer in the local cache with a
+--limit. If no ttl set will be 0
+function _M:save_issuer_in_cache(issuer, config, ttl)
+  local expires = tonumber(ttl) or 0
+
+  if not self.cache then
+    return nil
+  end
+  self.cache:set(issuer, config, expires)
+  ngx.log(ngx.DEBUG, "OIDC configuration for issuer='"..issuer.."' stored in cache for "..expires.." seconds")
 end
 
 --- Fetch and return OIDC configuration from well known endpoint <issuer>/.well-known/openid-configuration
@@ -75,12 +112,10 @@ function _M:openid_configuration(issuer)
     end
 
     local config = decode_json(res)
-
     if not config then
         ngx.log(oidc_log_level, 'invalid OIDC Provider, expected application/json got:  ', res.headers.content_type, ' body: ', res.body)
         return nil, 'invalid JSON'
     end
-
     return config
 end
 
@@ -115,18 +150,35 @@ end
 
 --- Fetch whole OIDC configuration through OIDC Discovery.
 -- @tparam string issuer URL to the Issuer (without the .well-known/openid-configuration)
+-- @tparam integer ttl for storing the issuer configuration in the local cache.
 -- @treturn table
-function _M:call(issuer)
-    local http_client = self.http_client
+function _M:call(issuer, ttl)
+  if not issuer then
+    return nil
+  end
 
-    if not http_client then
-        return nil, 'not initialized'
-    end
+  local http_client = self.http_client
 
-    local config, err = _M.openid_configuration(self, issuer)
-    if not config then return nil, err end
+  if not http_client then
+    return nil, 'not initialized'
+  end
 
-    return { config = config, issuer = config.issuer, keys = _M.jwks(self, config) }
+  local cached = self:issuer_in_cache(issuer)
+  if cached then
+    ngx.log(ngx.DEBUG, "retrieve OIDC configuration for issuer='"..issuer.."' from cache")
+    return cached
+  end
+
+  local config, err = _M.openid_configuration(self, issuer)
+  if not config then return nil, err end
+
+  local result =  {
+    config = config,
+    issuer = config.issuer,
+    keys = _M.jwks(self, config)
+  }
+  self:save_issuer_in_cache(issuer, result, tonumber(ttl))
+  return result
 end
 
 return _M